requirements.txt
python-telegram-bot==21.7
aiohttp>=3.9.0
pymongo>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
